requirements.txt
opencv-python==4.12.0.88
imutils==0.5.4
numpy==2.2.6
pyautogui==0.9.54
scipy==1.16.1
pillow==11.3.0
